Enum solicit::client::async::WorkItem

enum WorkItem {
    Request(AsyncRequest),
    HandleFrame,
    SendData,
    NewClient,
    ClientLeft,
}

An enum representing the types of work that the ClientService can perform from within its run_once method.

Variants

Request

Queue a new request to the HTTP/2 connection.

HandleFrame

Trigger a new handle_next_frame. The work item should be queued only when there is a frame to be handled to avoid blocking the run_once call.

SendData

Trigger a new send_next_data operation.

NewClient

Signals to the service that a new client is connected. Helps it keep track of whether there are clients that would expect a response.

ClientLeft

Signals to the service that a client has disconnected. Helps it keep track of whether there are clients that would expect a response.