Struct solicit::client::async::ChannelFrameReceiver

struct ChannelFrameReceiver<R> where R: ReceiveFrame {
    tx: Sender<HttpFrame>,
    inner: R,
}

A struct that buffers HttpFrames read by the wrapped ReceiveFrame instance in an internal mpsc channel. The reads from the wrapped ReceiveFrame instance are triggered by calls to the read_next method.

Additionally, it provides a ChannelFrameReceiverHandle instance that implements the ReceiveFrame trait, such that it pops the next available frame from the internal channel. If there are no available frames, it will block, so care must be taken to trigger the connection's frame handling only when there are buffered frames, if it is not to block.

As such, this is a convenience struct that makes it possible to provide non-blocking reads from within HttpConnections, while handling the actual reads using a ReceiveFrame implementation that can block. (Predicated on triggering a single frame handle operation on the connection for each successfully executed read_next.)

Fields

tx

The sender side of the channel. Buffers the frames read by the wrapped ReceiveFrame instance for future consumation by the associated ChannelFrameReceiverHandle.

inner

The ReceiveFrame instance that performs the actual reading of the frame, used from within the read_next method.

Methods

impl<R> ChannelFrameReceiver<R> where R: ReceiveFrame

fn new(inner: R) -> (ChannelFrameReceiver<R>, ChannelFrameReceiverHandle)

Creates a new ChannelFrameReceiver, as well as the associated ChannelFrameReceiverHandle.

fn read_next(&mut self) -> HttpResult<()>

Performs a recv_frame operation on the wrapped ReceiveFrame instance, possibly blocking the thread in the process, depending on the implementation of the trait. Once a frame is returned, it will buffer it within the internal channel.