Struct solicit::client::async::ChannelFrameReceiverHandle

struct ChannelFrameReceiverHandle {
    rx: Receiver<HttpFrame>,
}

A handle to the ChannelFrameReceiver and an implementation of the ReceiveFrame trait. It simply pops the next frame from the internal channel that buffers the frames read by the ReceiveFrame instance wrapped by the associated ChannelFrameReceiver. If there are no frames currently buffered, it blocks until there is one. Therefore, the handle_next_frame method of the HttpConnection that relies on the IO provided by this ReceiveFrame implementation should be triggered only when sure that there are buffered frames, if blocking handles are to be avoided.

Fields

rx

The receiver end of the channel that buffers the received frames.