Struct solicit::client::async::ChannelFrameSender

struct ChannelFrameSender<S> where S: SendFrame {
    rx: Receiver<RawFrame>,
    inner: S,
}

A struct that buffers RawFrames in an internal mpsc channel and sends them using the wrapped SendFrame instance when the send_next method is called.

Additionally, it provides a ChannelFrameSenderHandle instance that implements the SendFrame trait and as such can be passed to the HttpConnection. This handler simply queues the frame into the internal channel, without ever blocking.

As such, this is a convenience struct that makes it possible to provide non-blocking writes from within HttpConnections, while handling the actual writes using a SendFrame implementation that will block until the frame is sent on a separate thread.

Fields

rx

The receiving end of the channel. Buffers the frames that are to be sent.

inner

The SendFrame instance that will perform the actual writes from within the send_next method.

Methods

impl<S> ChannelFrameSender<S> where S: SendFrame

fn new(inner: S) -> (ChannelFrameSender<S>, ChannelFrameSenderHandle)

Creates a new ChannelFrameSender that will use the provided SendFrame instance within the send_next method in order to perform the final send to the remote peer. The ChannelFrameSenderHandle that is returned can be used to queue frames for sending from within HttpConnections, as it implements the SendFrame trait.

fn send_next(&mut self) -> HttpResult<()>

Performs the send of the next frame that is buffered in the internal channel of the struct.

If there is no frame in the channel, it will block until there is one there.

If the channel becomes disconnected from all senders, indicating that all handles to the sender have been dropped, the mehod will return an error.