Struct solicit::client::async::ChannelFrameSenderHandle

struct ChannelFrameSenderHandle {
    tx: Sender<RawFrame>,
}

A handle to the ChannelFrameSender and an implementation of the SendFrame trait. It simply queues the given frames into the send queue of the ChannelFrameSender without ever blocking. (Except possibly to allocate some memory, as per the mpsc::channel specification.)

Fields
tx: Sender<RawFrame>

The sender side of the channel that buffers the frames to be written. Allows the handle to queue the frame for future writing without blocking on the IO.