Struct solicit::client::async::ClientService

struct ClientService {
    next_stream_id: StreamId,
    outstanding_reqs: u32,
    limit: u32,
    conn: ClientConnection<ChannelFrameSenderHandle, ChannelFrameReceiverHandle>,
    chans: HashMap<StreamId, Sender<Response>>,
    work_queue: Receiver<WorkItem>,
    request_queue: Vec<AsyncRequest>,
    client_count: i32,
    host: Vec<u8>,
    initialized: bool,
}

An internal struct encapsulating a service that lets multiple clients issue concurrent requests to the same HTTP/2 connection.

The service maintains an internal queue of WorkItems that indicate what the operations that it should perform. The next operation from the queue is performed on each run_once method call.

It handles issuing new requests (corresponding to WorkItem::Request work item), handling the next received frame (when indicated by the WorkItem::HandleFrame), and tracks the number of connected clients (run_once returns an error once there are no more clients connected to the service).

If there is no work in the queue, the run_once method blocks.

Essentially, this represents a simplified event loop that handles events queued on the work queue (blocking to wait for new work when none is available; does not spin). Therefore, the user of the ClientService needs to provide a dedicated thread in which to run the run_once event loop handler.

Additionally, the client needs to make sure to perform the actual socket IO (which is fully blocking, without even timeout support currently in Rust) in threads dedicated for that, by calling the send_next or read_next methods of the ChannelFrameSender or ChannelFrameReceiver, which are returned from the ClientService constructor.

TODO: Technically, the run_once method could take a WorkItem, so a single event loop could dispatch work items to a corresponding service, removing the need for the thread-per-service requirement. However, at that point we're nearing a reimplementation of a real event loop, which is slightly out of scope of the solicit library, as imagined; the async client is (for now) supposed to be a proof-of-concept implementation of a high-level async/concurrent HTTP/2 client.

Fields
next_stream_id: StreamId

The ID that will be assigned to the next client-initiated stream.

outstanding_reqs: u32

The number of requests that have been sent, but are yet unanswered.

limit: u32

The limit to the number of requests that can be pending (unanswered, but sent).

conn: ClientConnection<ChannelFrameSenderHandle, ChannelFrameReceiverHandle>

The connection that is used for underlying HTTP/2 communication.

chans: HashMap<StreamId, Sender<Response>>

A mapping of stream IDs to the sender side of a channel that is expecting a response to the request that is to arrive on that stream.

work_queue: Receiver<WorkItem>

The receiver end of a channel to which work items for the service are queued. Work items include the variants of the WorkItem enum.

request_queue: Vec<AsyncRequest>

The queue of AsyncRequests that haven't yet been sent to the server.

client_count: i32

Tracks the number of currently connected clients -- once it reaches 0, the run_once method returns an error.

host: Vec<u8>

The name of the host the connection is established to.

initialized: bool

Whether the connection has already been initialized.

Methods

impl ClientService

fn new<S>(client_stream: ClientStream<S>) -> Option<Service<S>> where S: TransportStream

Creates a new ClientService that will use the provided ClientStream for its underlying network communication. A handle is returned for both the read, as well as the write end of the socket that allows the client that creates the ClientService to perform the blocking IO without influencing the ClientService (i.e. without having its run_once method block).

Returns

Returns all the relevant components of the newly created ClientService:

The ClientService itself -- processes events (WorkItems) on each run_once call.

The sender-side of the work queue -- allows WorkItems to be queued into the ClientService's simplified event loop.

The ChannelFrameReceiver -- the instance that wraps the actual socket that performs the blocking read IO. Allows the caller to block on the IO in a customized manner (e.g. in a separate dedicated thread).

The ChannelFrameSender -- the instance that wraps the actual socket that performs the blocking write IO. Allows the caller to block on the IO in a customized manner (e.g. in a separate thread).

If no HTTP/2 connection can be established to the given host on the given port, returns None.

fn run_once(&mut self) -> Result<(), ClientServiceErr>

Performs one iteration of the service.

One iteration corresponds to running the next WorkItem that the service has queued in its work_queue. Essentially, this is a poor-man's event loop implementation. If there is no work queued for the service, it will block, until there is. As such, embedding calls to this method into a real event loop should not be done.

For WorkItem::Request work items, the service will queue the received AsyncRequest for sending. It will also attempt to queue it for transmission to the server, unless the concurrent requests limit has been exceeded, in which case the request is kept in an internal FIFO queue and will be sent when its time comes.

For WorkItem::HandleFrame work items, the service will perform a single handle_next_frame call on its underlying ClientConnection instance. Since the item is queued only when the connection actually has frames to process, this call will never block. If a response got finalized by the handling of the frame, it is shipped to the channel that expects it and a new request from the request queue sent.

Returns

On a successful pass, the function returns an Ok(()).

The Err response is returned when there are no more responses to be received and there are no more clients connected to the service (and thus no more requests could ever be issued by the instance). This corresponds to the ClientServiceErr::Done variant.

Any HTTP/2 error is propagated (wrapped into a ClientServiceErr::Http variant).

fn handle_frame(&mut self) -> Result<(), ClientServiceErr>

A private convenience method that performs the handling of the next received frame.

It calls the underlying connection's handle_next_frame method and then inspects the changes made to the session, notifying clients of completed requests or queueing new ones, if available.

fn send_request(&mut self, async_req: AsyncRequest)

Internal helper method. Sends a request to the server based on the parameters given in the AsyncRequest. It blocks until the request is fully transmitted to the server.

fn create_request(&mut self, async_req: AsyncRequest) -> (RequestStream<DefaultStream>, Sender<Response>)

Internal helper method. Creates a new RequestStream instance based on the given parameters. Such a RequestStream instance is ready to be passed to the connection for transmission to the server (i.e. start_request). Also returns the sender end of the channel to which the response is to be transmitted, once received.

fn send_response(&mut self, stream: DefaultStream)

Internal helper method. Sends a response assembled from the given stream to the corresponding channel that is waiting for the response.

The given stream instance is consumed by this method.

fn handle_closed(&mut self)

Internal helper method. Handles all closed streams by sending appropriate notifications to waiting channels.

For now, the channels are all given a Response, even though the stream might end up being closed by the server with an error.

fn queue_next_request(&mut self)

Internal helper method. If there are yet unsent requests queued by a client to the service and the service has not exceeded the limit of concurrent requests that it is allowed to issue, it sends a single new request to the server. Blocks until this request is sent.