Function solicit::http::client::write_preface

pub fn write_preface<W: Write>(stream: &mut W) -> Result<(), Error>

Writes the client preface to the underlying HTTP/2 connection.

According to the HTTP/2 spec, a client preface is first a specific sequence of octets, followed by a settings frame.

Returns

Any error raised by the underlying connection is propagated.