Struct solicit::http::client::CleartextConnector

pub struct CleartextConnector<'a> {
    pub host: &'a str,
    pub port: u16,
}

A struct that establishes a cleartext TCP connection that can be used by an HTTP/2 connection. Defaults to using port 80.

It assumes that the connection is based on prior knowledge of the server's support for HTTP/2.

More information in the spec

Fields
host: &'a str

The host to which the connection should be established

port: u16

The port on which the connection should be established

Methods

impl<'a> CleartextConnector<'a>

fn new(host: &'a str) -> CleartextConnector

Creates a new CleartextConnector that will attempt to establish a connection to the given host on port 80.

fn with_port(host: &'a str, port: u16) -> CleartextConnector

Creates a new CleartextConnector that will attempt to establish a connection to the given host on the given port.