Struct solicit::http::client::RequestStream

pub struct RequestStream<S> where
    S: Stream,  {
    pub headers: Vec<Header>,
    pub stream: S,
}

A struct representing a request stream. It provides the headers that are to be sent when initiating the request, as well as a Stream instance that handles the received response and provides the request body.

Fields
headers: Vec<Header>

The list of headers that will be sent with the request.

stream: S

The underlying Stream instance, which will handle the response, as well as optionally provide the body of the request.