Enum solicit::http::connection::TlsConnectError

pub enum TlsConnectError {
    IoError(Error),
    SslError(SslError),
    Http2NotSupported(SslStream<TcpStream>),
}

An enum representing possible errors that can arise when trying to establish an HTTP/2 connection over TLS.

Variants

IoError

The variant corresponds to the underlying raw TCP connection returning an error.

SslError

The variant corresponds to the TLS negotiation returning an error.

Http2NotSupported

The variant corresponds to the case when the TLS connection is established, but the application protocol that was negotiated didn't end up being HTTP/2. It wraps the established SSL stream in order to allow the client to decide what to do with it (and the application protocol that was chosen).