Struct solicit::http::connection::ClientConnection

pub struct ClientConnection<TS, S> where TS: TransportStream, S: Session {
    pub session: S,
    // some fields omitted
}

A struct implementing the client side of an HTTP/2 connection.

It builds on top of an HttpConnection and provides additional methods that are only used by clients.

Fields

session

The Session associated with this connection. It is essentially a set of callbacks that are triggered by the connection when different states in the HTTP/2 communication arise.

Methods

impl<TS, S> ClientConnection<TS, S> where TS: TransportStream, S: Session

fn with_connection(conn: HttpConnection<TS>, session: S) -> ClientConnection<TS, S>

Creates a new ClientConnection that will use the given HttpConnection for all its underlying HTTP/2 communication.

fn host(&self) -> &str

Returns the host to which the underlying HttpConnection is established.

fn scheme(&self) -> HttpScheme

Returns the scheme of the underlying HttpConnection.

fn init(&mut self) -> HttpResult<()>

Performs the initialization of the ClientConnection.

Sends the client preface, followed by validating the receipt of the server preface.

fn send_request(&mut self, req: Request) -> HttpResult<()>

A method that sends the given Request to the server.

The method blocks until the entire request has been sent.

All errors are propagated.

Note

Request body is ignored for now.

fn handle_next_frame(&mut self) -> HttpResult<()>

Fully handle the next incoming frame, blocking to read it from the underlying transport stream if not available yet.

All communication errors are propagated.