Struct solicit::http::connection::HttpConnection

pub struct HttpConnection<S> where S: TransportStream {
    pub scheme: HttpScheme,
    pub host: String,
    // some fields omitted
}

The struct implements the HTTP/2 connection level logic.

It provides an API for writing and reading HTTP/2 frames. It also takes care to validate the received frames.

Fields

scheme

The scheme of the connection

host

The host to which the connection is established

Methods

impl<S> HttpConnection<S> where S: TransportStream

fn new<'a>(stream: S, scheme: HttpScheme, host: Cow<'a, str>) -> HttpConnection<S>

Creates a new HttpConnection that will use the given stream as its underlying transport layer.

The host to which the connection is established, as well as the connection scheme are provided.

fn send_frame<F: Frame>(&mut self, frame: F) -> HttpResult<()>

Sends the given frame to the peer.

Returns

Any IO errors raised by the underlying transport layer are wrapped in a HttpError::IoError variant and propagated upwards.

If the frame is successfully written, returns a unit Ok (Ok(())).

fn recv_frame(&mut self) -> HttpResult<HttpFrame>

Reads a new frame from the transport layer.

Returns

Any IO errors raised by the underlying transport layer are wrapped in a HttpError::IoError variant and propagated upwards.

If the frame type is unknown the HttpError::UnknownFrameType variant is returned.

If the frame type is recognized, but the frame cannot be successfully decoded, the HttpError::InvalidFrame variant is returned. For now, invalid frames are not further handled by informing the peer (e.g. sending PROTOCOL_ERROR) nor can the exact reason behind failing to decode the frame be extracted.

If a frame is successfully read and parsed, returns the frame wrapped in the appropriate variant of the HttpFrame enum.