Struct solicit::http::connection::TlsConnector

pub struct TlsConnector<'a, 'ctx> {
    pub host: &'a str,
    // some fields omitted
}

A struct implementing the functionality of establishing an HTTP/2 connection over a TLS-backed stream. The protocol negotiation also takes place within the TLS negotiation.

Fields

host

Methods

impl<'a, 'ctx> TlsConnector<'a, 'ctx>

fn new<P: AsRef<Path>>(host: &'a str, ca_file_path: &'ctx P) -> TlsConnector<'a, 'ctx>

Creates a new TlsConnector that will create a new SslContext before trying to establish the TLS connection. The path to the CA file that the context will use needs to be provided.

fn with_context(host: &'a str, context: &'ctx SslContext) -> TlsConnector<'a, 'ctx>

Creates a new TlsConnector that will use the provided context to create the SslStream that will back the HTTP/2 connection.

fn build_default_context(ca_file_path: &Path) -> Result<SslContext, TlsConnectError>

Builds up a default SslContext instance wth TLS settings that the HTTP/2 spec mandates. The path to the CA file needs to be provided.