Trait solicit::http::connection::HttpConnect

pub trait HttpConnect {
    type Stream: TransportStream;
    type Err;

    fn connect(self) -> Result<HttpConnection<Self::Stream>, Self::Err>;
}

A trait that can be implemented by structs that want to provide the functionality of establishing HTTP/2 connections.

Associated Types

type Stream: TransportStream

The type of the underlying transport stream that the HttpConnections produced by this HttpConnect implementation will be based on.

type Err

The type of the error that can be produced by trying to establish the connection (i.e. calling the connect method).

Required Methods

fn connect(self) -> Result<HttpConnection<Self::Stream>, Self::Err>

Establishes an HTTP/2 connection...