Enum solicit::http::ErrorCode

pub enum ErrorCode {
    NoError,
    ProtocolError,
    InternalError,
    FlowControlError,
    SettingsTimeout,
    StreamClosed,
    FrameSizeError,
    RefusedStream,
    Cancel,
    CompressionError,
    ConnectError,
    EnhanceYourCalm,
    InadequateSecurity,
    Http11Required,
}

The enum represents an error code that are used in RST_STREAM and GOAWAY frames. These are defined in Section 7 of the HTTP/2 spec.

Variants
NoError

The associated condition is not a result of an error. For example, a GOAWAY might include this code to indicate graceful shutdown of a connection.

ProtocolError

The endpoint detected an unspecific protocol error. This error is for use when a more specific error code is not available.

InternalError

The endpoint encountered an unexpected internal error.

FlowControlError

The endpoint detected that its peer violated the flow-control protocol.

SettingsTimeout

The endpoint sent a SETTINGS frame but did not receive a response in a timely manner. See Section 6.5.3 ("Settings Synchronization").

StreamClosed

The endpoint received a frame after a stream was half-closed.

FrameSizeError

The endpoint received a frame with an invalid size.

RefusedStream

The endpoint refused the stream prior to performing any application processing (see Section 8.1.4 for details).

Cancel

Used by the endpoint to indicate that the stream is no longer needed.

CompressionError

The endpoint is unable to maintain the header compression context for the connection.

ConnectError

The connection established in response to a CONNECT request (Section 8.3) was reset or abnormally closed.

EnhanceYourCalm

The endpoint detected that its peer is exhibiting a behavior that might be generating excessive load.

InadequateSecurity

The underlying transport has properties that do not meet minimum security requirements (see Section 9.2).

Http11Required

The endpoint requires that HTTP/1.1 be used instead of HTTP/2.