Struct solicit::http::frame::goaway::GoawayFrame

pub struct GoawayFrame<'a> { /* fields omitted */ }

The struct represents the GOAWAY HTTP/2 frame.

Implementations

impl<'a> GoawayFrame<'a>

pub fn new(last_stream_id: StreamId, error_code: ErrorCode) -> Self

Create a new GOAWAY frame with the given error code and no debug data.

pub fn with_debug_data(
last_stream_id: StreamId,
raw_error: u32,
debug_data: &'a [u8]
) -> Self

Create a new GOAWAY frame with the given parts.

pub fn error_code(&self) -> ErrorCode

Returns the interpreted error code of the frame. Any unknown error codes are mapped into the InternalError variant of the enum.

pub fn raw_error_code(&self) -> u32

Returns the original raw error code of the frame. If the code is unknown, it will not be changed.

pub fn last_stream_id(&self) -> StreamId

Returns the associated last stream ID.

pub fn debug_data(&self) -> Option<&[u8]>

Returns the debug data associated with the frame.

pub fn payload_len(&self) -> u32

Returns the total length of the frame's payload, including any debug data.