Module solicit::http::frame::headers

The module contains the implementation of the HEADERS frame and associated flags.

Structs

HeadersFrame
A struct representing the HEADERS frames of HTTP/2, as defined in the HTTP/2 spec, section 6.2.

StreamDependency
The struct represents the dependency information that can be attached to a stream and sent within a HEADERS frame (one with the Priority flag set).

Enums

HeadersFlag
An enum representing the flags that a HeadersFrame can have. The integer representation associated to each variant is that flag's bitmask.