Module solicit::http::frame::settings

The module contains the implementation of the SETTINGS frame and associated flags.

Structs

SettingsFrame
A struct representing the SETTINGS frames of HTTP/2, as defined in the HTTP/2 spec, section 6.5.

Enums

HttpSetting
An enum that lists all valid settings that can be sent in a SETTINGS frame.

SettingsFlag
An enum representing the flags that a SettingsFrame can have. The integer representation associated to each variant is that flag's bitmask.