Enum solicit::http::frame::settings::HttpSetting

pub enum HttpSetting {
    HeaderTableSize(u32),
    EnablePush(u32),
    MaxConcurrentStreams(u32),
    InitialWindowSize(u32),
    MaxFrameSize(u32),
    MaxHeaderListSize(u32),
}

An enum that lists all valid settings that can be sent in a SETTINGS frame.

Each setting has a value that is a 32 bit unsigned integer (6.5.1.).

Variants
HeaderTableSize(u32)EnablePush(u32)MaxConcurrentStreams(u32)InitialWindowSize(u32)MaxFrameSize(u32)MaxHeaderListSize(u32)
Methods

impl HttpSetting

fn from_id(id: u16, val: u32) -> Option<HttpSetting>

Creates a new HttpSetting with the correct variant corresponding to the given setting id, based on the settings IDs defined in section 6.5.2.

fn get_id(&self) -> u16

Returns the setting ID as an unsigned 16 bit integer, as defined in section 6.5.2.

fn get_val(&self) -> u32

Gets the setting value by unpacking it from the wrapped u32.