Enum solicit::http::session::StreamDataError

pub enum StreamDataError {
    Closed,
    Other(Box<Error + Send + Sync>),
}

The enum represents errors that can be returned from the Stream::get_data_chunk method.

Variants
Closed

Indicates that the stream cannot provide any data, since it is closed for further writes from the peer's side.

Other(Box<Error + Send + Sync>)

A different error while trying to obtain the data chunk. Wraps a boxed Error impl.