Struct solicit::http::session::DefaultSession

pub struct DefaultSession<S = DefaultStream> where S: Stream {
    // some fields omitted
}

A simple implementation of the Session trait.

Keeps track of which streams are currently active by holding a HashMap of stream IDs to Stream instances. Callbacks delegate to the corresponding stream instance, after validating the received stream ID.

The purpose of the type is to make it easier for client implementations to only handle stream-level events by providing a Stream implementation, instead of having to implement the entire session management (tracking active streams, etc.).

For example, by varying the Stream implementation it is easy to implement a client that streams responses directly into a file on the local file system, instead of keeping it in memory (like the DefaultStream does), without having to change any HTTP/2-specific logic.

Methods

impl<S> DefaultSession<S> where S: Stream

fn new() -> DefaultSession<S>

Returns a new DefaultSession with no active streams.

fn get_stream(&self, stream_id: StreamId) -> Option<&S>

Returns a reference to a stream with the given ID, if such a stream is found in the DefaultSession.

fn new_stream(&mut self, stream_id: StreamId)

Creates a new stream with the given ID in the session.

fn get_closed(&mut self) -> Vec<S>

Returns all streams that are closed and tracked by the session.

The streams are moved out of the session.