Struct solicit::http::Response

pub struct Response {
    pub stream_id: StreamId,
    pub headers: Vec<Header>,
    pub body: Vec<u8>,
}

A struct representing the full raw response received on an HTTP/2 connection.

The full body of the response is included, regardless how large it may be. The headers contain both the meta-headers, as well as the actual headers.

Fields
stream_id: StreamId

The ID of the stream to which the response is associated. HTTP/1.1 does not really have an equivalent to this.

headers: Vec<Header>

Exposes all the raw response headers, including the meta-headers. (For now the only meta header allowed in HTTP/2 responses is the :status.)

body: Vec<u8>

The full body of the response as an uninterpreted sequence of bytes.

Methods

impl Response

fn new(stream_id: StreamId, headers: Vec<Header>, body: Vec<u8>) -> Response

Creates a new Response with all the components already provided.

fn status_code(&self) -> HttpResult<u16>

Gets the response status code from the pseudo-header. If the response does not contain the response as the first pseuo-header, an error is returned as such a response is malformed.