Struct solicit::http::transport::TransportReceiveFrame

pub struct TransportReceiveFrame<'a, TS> where TS: TransportStream + 'a { /* fields omitted */ }

The struct is a an implementation of the ReceiveFrame trait that wraps an existing TransportStream and uses it to provide HTTP/2 frames, when asked for one, by reading from the stream. The implementation always allocates a new buffer on the heap for every incoming frame.

Methods

impl<'a, TS> TransportReceiveFrame<'a, TS> where TS: TransportStream

fn new(ts: &'a mut TS) -> TransportReceiveFrame<'a, TS>

Create a new TransportReceiveFrame that will use the given TransportStream for reading the frame.