Module solicit::http::client

The module contains a number of reusable components for implementing the client side of an HTTP/2 connection.

Structs

CleartextConnectError
A newtype wrapping the io::Error, as it occurs when attempting to establish an HTTP/2 connection over cleartext TCP (with prior knowledge).

CleartextConnector
A struct that establishes a cleartext TCP connection that can be used by an HTTP/2 connection. Defaults to using port 80.

ClientConnection
The struct extends the HttpConnection API with client-specific methods (such as start_request) and wires the HttpConnection to the client Session callbacks.

ClientSession
An implementation of the Session trait specific to handling client HTTP/2 connections.

ClientStream
A convenience wrapper type that represents an established client network transport stream. It wraps the stream itself, the scheme of the protocol to be used, and the remote host name.

RequestStream
A struct representing a request stream. It provides the headers that are to be sent when initiating the request, as well as a Stream instance that handles the received response and provides the request body.

Traits

HttpConnect
A trait that can be implemented by structs that want to provide the functionality of establishing network connections for use by HTTP/2 connections.

HttpConnectError
A marker trait for errors raised by attempting to establish an HTTP/2 connection.

Functions

write_preface
Writes the client preface to the underlying HTTP/2 connection.