Struct solicit::http::client::ClientConnection

pub struct ClientConnection<State = DefaultSessionState<ClientMarker, DefaultStream>> where State: SessionState {
    pub state: State,
    // some fields omitted
}

The struct extends the HttpConnection API with client-specific methods (such as start_request) and wires the HttpConnection to the client Session callbacks.

Fields

state

The state of the session associated to this client connection. Maintains the status of the connection streams.

Methods

impl<State> ClientConnection<State> where State: SessionState

fn with_connection(conn: HttpConnection, state: State) -> ClientConnection<State>

Creates a new ClientConnection that will use the given HttpConnection for all its underlying HTTP/2 communication.

The given state instance will handle the maintenance of the session's state.

fn scheme(&self) -> HttpScheme

Returns the scheme of the underlying HttpConnection.

fn expect_settings<Recv: ReceiveFrame, Sender: SendFrame>(&mut self, rx: &mut Recv, tx: &mut Sender) -> HttpResult<()>

Handles the next frame provided by the given frame receiver and expects it to be a SETTINGS frame. If it is not, it returns an error.

The method is a convenience method that can be used during the initialization of the connection, as the first frame that any peer is allowed to send is an initial settings frame.

fn start_request<S: SendFrame>(&mut self, req: RequestStream<State::Stream>, sender: &mut S) -> HttpResult<StreamId>

Starts a new request based on the given RequestStream.

For now it does not perform any validation whether the given RequestStream is valid.

fn handle_next_frame<Recv: ReceiveFrame, Sender: SendFrame>(&mut self, rx: &mut Recv, tx: &mut Sender) -> HttpResult<()>

Fully handles the next incoming frame provided by the given ReceiveFrame instance. Handling a frame may cause changes to the session state exposed by the ClientConnection.

fn send_next_data<S: SendFrame>(&mut self, sender: &mut S) -> HttpResult<SendStatus>

Queues a new DATA frame onto the underlying SendFrame.

Currently, no prioritization of streams is taken into account and which stream's data is queued cannot be relied on.