Struct solicit::http::client::ClientSession

pub struct ClientSession<'a, State, S> where State: SessionState + 'a, S: SendFrame + 'a { /* fields omitted */ }

An implementation of the Session trait specific to handling client HTTP/2 connections.

While handling the events signaled by the HttpConnection, the struct will modify the given session state appropriately.

The purpose of the type is to make it easier for client implementations to only handle stream-level events by providing a Stream implementation, instead of having to implement all session management callbacks.

For example, by varying the Stream implementation it is easy to implement a client that streams responses directly into a file on the local file system, instead of keeping it in memory (like the DefaultStream does), without having to change any HTTP/2-specific logic.

Methods

impl<'a, State, S> ClientSession<'a, State, S> where State: SessionState + 'a, S: SendFrame + 'a

fn new(state: &'a mut State, sender: &'a mut S) -> ClientSession<'a, State, S>

Returns a new ClientSession associated to the given state.