Struct solicit::http::connection::DataChunk

pub struct DataChunk<'a> {
    pub data: Cow<'a, [u8]>,
    pub stream_id: StreamId,
    pub end_stream: EndStream,
}

The struct represents a chunk of data that should be sent to the peer on a particular stream.

Fields
data: Cow<'a, [u8]>

The data that should be sent.

stream_id: StreamId

The ID of the stream on which the data should be sent.

end_stream: EndStream

Whether the data chunk will also end the stream.

Methods

impl<'a> DataChunk<'a>

fn new(data: Cow<'a, [u8]>, stream_id: StreamId, end_stream: EndStream) -> DataChunk<'a>

Creates a new DataChunk.

Note: IntoCow is unstable and there's no implementation of Into<Cow<'a, [u8]>> for the fundamental types, making this a bit of a clunky API. Once such an Into impl is added, this can be made generic over the trait for some ergonomic improvements.

fn new_borrowed<D: Borrow<&'a [u8]>>(data: D, stream_id: StreamId, end_stream: EndStream) -> DataChunk<'a>

Creates a new DataChunk from a borrowed slice. This method should become obsolete if we can take an Into<Cow<_, _>> without using unstable features.