Struct solicit::http::connection::HttpConnectionSender

pub struct HttpConnectionSender<'a, S> where S: SendFrame + 'a { /* fields omitted */ }

The struct represents an HttpConnection that has been bound to a SendFrame reference, allowing it to send frames. It exposes convenience methods for various send operations that can be invoked on the underlying stream. The methods prepare the appropriate frames and queue their sending on the referenced SendFrame instance.

The only way for clients to obtain an HttpConnectionSender is to invoke the HttpConnection::sender method and provide it a reference to the SendFrame that should be used.

Methods

impl<'a, S> HttpConnectionSender<'a, S> where S: SendFrame + 'a

fn rst_stream(&mut self, id: StreamId, code: ErrorCode) -> HttpResult<()>

Send a RST_STREAM frame for the given frame id

fn send_settings_ack(&mut self) -> HttpResult<()>

Sends a SETTINGS acknowledge frame to the peer.

fn send_ping_ack(&mut self, bytes: u64) -> HttpResult<()>

Sends a PING ack

fn send_ping(&mut self, bytes: u64) -> HttpResult<()>

Sends a PING request

fn send_headers<'n, 'v, H: Into<Vec<Header<'n, 'v>>>>(&mut self,
headers: H,
stream_id: StreamId,
end_stream: EndStream)
-> HttpResult<()>

A helper function that inserts the frames required to send the given headers onto the SendFrame stream.

The HttpConnection performs the HPACK encoding of the header block using an internal encoder.

Parameters

headers - a headers list that should be sent.

stream_id - the ID of the stream on which the headers will be sent. The connection performs no checks as to whether the stream is a valid identifier.

end_stream - whether the stream should be closed from the peer's side immediately after sending the headers

fn send_data(&mut self, chunk: DataChunk) -> HttpResult<()>

A helper function that inserts a frame representing the given data into the SendFrame stream. In doing so, the connection's outbound flow control window is adjusted appropriately.

fn send_next_data<P: DataPrioritizer>(&mut self,
prioritizer: &mut P)
-> HttpResult<SendStatus>

Sends the chunk of data provided by the given DataPrioritizer.

Returns

Returns the status of the operation. If the provider does not currently have any data that could be sent, returns SendStatus::Nothing. If any data is sent, returns SendStatus::Sent.