Module solicit::http::frame

The module contains the implementation of HTTP/2 frames.

Re-exports

pub use self::data::DataFlag;

pub use self::data::DataFrame;

pub use self::headers::HeadersFlag;

pub use self::headers::HeadersFrame;

pub use self::settings::SettingsFlag;

pub use self::settings::SettingsFrame;

pub use self::settings::HttpSetting;

Modules

data
The module contains the implementation of the DATA frame and associated flags.

headers
The module contains the implementation of the HEADERS frame and associated flags.

settings
The module contains the implementation of the SETTINGS frame and associated flags.

Structs

RawFrame
A struct that defines the format of the raw HTTP/2 frame, i.e. the frame as it is read from the wire.

Traits

Flag
A trait that all HTTP/2 frame header flags need to implement.

Frame
A trait that all HTTP/2 frame structs need to implement.

Functions

pack_header
Constructs a buffer of 9 bytes that represents the given FrameHeader.

unpack_header
Deconstructs a FrameHeader into its corresponding 4 components, represented as a 4-tuple: (length, frame_type, flags, stream_id).

Type Definitions

FrameHeader
An alias for the 4-tuple representing the components of each HTTP/2 frame header.

FrameHeaderBuffer
An alias for the 9-byte buffer that each HTTP/2 frame header must be stored in.