Struct solicit::http::frame::data::DataFrame

pub struct DataFrame {
    pub data: Vec<u8>,
    // some fields omitted
}

A struct representing the DATA frames of HTTP/2, as defined in the HTTP/2 spec, section 6.1.

Fields
data: Vec<u8>

The data found in the frame as an opaque byte sequence. It never includes padding bytes.

Methods

impl DataFrame

fn new(stream_id: StreamId) -> DataFrame

Creates a new empty DataFrame, associated to the stream with the given ID.

fn is_padded(&self) -> bool

Returns true if the DATA frame is padded, otherwise false.

fn is_end_of_stream(&self) -> bool

Returns whther this frame ends the stream it is associated with.

fn set_padding(&mut self, pad_len: u8)

Sets the number of bytes that should be used as padding for this frame.