Struct solicit::http::frame::headers::HeadersFrame

pub struct HeadersFrame {
    pub header_fragment: Vec<u8>,
    pub stream_id: StreamId,
    pub stream_dep: Option<StreamDependency>,
    pub padding_len: Option<u8>,
    // some fields omitted
}

A struct representing the HEADERS frames of HTTP/2, as defined in the HTTP/2 spec, section 6.2.

Fields
header_fragment: Vec<u8>

The header fragment bytes stored within the frame.

stream_id: StreamId

The ID of the stream with which this frame is associated

stream_dep: Option<StreamDependency>

The stream dependency information, if any.

padding_len: Option<u8>

The length of the padding, if any.

Methods

impl HeadersFrame

fn new(fragment: Vec<u8>, stream_id: StreamId) -> HeadersFrame

Creates a new HeadersFrame with the given header fragment and stream ID. No padding, no stream dependency, and no flags are set.

fn with_dependency(fragment: Vec<u8>, stream_id: StreamId, stream_dep: StreamDependency) -> HeadersFrame

Creates a new HeadersFrame with the given header fragment, stream ID and stream dependency information. No padding and no flags are set.

fn is_headers_end(&self) -> bool

Returns whether this frame ends the headers. If not, there MUST be a number of follow up CONTINUATION frames that send the rest of the header data.

fn is_end_of_stream(&self) -> bool

Returns whther this frame ends the stream it is associated with.

fn set_padding(&mut self, padding_len: u8)

Sets the padding length for the frame, as well as the corresponding Padded flag.