Struct solicit::http::frame::RawFrame

pub struct RawFrame<'a> { /* fields omitted */ }

A struct that defines the format of the raw HTTP/2 frame, i.e. the frame as it is read from the wire.

This format is defined in section 4.1. of the HTTP/2 spec.

The RawFrame struct simply stores the raw components of an HTTP/2 frame: its header and the payload as a sequence of bytes.

It does not try to interpret the payload bytes, nor do any validation in terms of its validity based on the frame type given in the header. It is simply a wrapper around the two parts of an HTTP/2 frame.

Methods

impl<'a> RawFrame<'a>

fn parse(buf: &'a [u8]) -> Option<RawFrame<'a>>

Parses a RawFrame from the bytes starting at the beginning of the given buffer.

Returns the None variant when it is not possible to parse a raw frame from the buffer (due to there not being enough bytes in the buffer). If the RawFrame is successfully parsed it returns the frame, borrowing a part of the original buffer. Therefore, this method makes no copies, nor does it perform any extra allocations.

Examples

use solicit::http::frame::RawFrame;

let buf = b"123";
// Not enough bytes for even the header of the frame
assert!(RawFrame::parse(&buf[..]).is_none());

use solicit::http::frame::RawFrame;

let buf = vec![0, 0, 1, 0, 0, 0, 0, 0, 0];
// Full header, but not enough bytes for the payload
assert!(RawFrame::parse(&buf[..]).is_none());

use solicit::http::frame::RawFrame;

let buf = vec![0, 0, 1, 0, 0, 0, 0, 0, 0, 1];
// A full frame is extracted, even if in this case the frame itself is not valid (a DATA
// frame associated to stream 0 is not valid in HTTP/2!). This, however, is not the
// responsibility of the RawFrame.
let frame = RawFrame::parse(&buf[..]).unwrap();
assert_eq!(frame.as_ref(), &buf[..]);

fn len(&self) -> usize

Returns the total length of the RawFrame, including both headers, as well as the entire payload.

fn serialize(&self) -> Vec<u8>

Returns a Vec of bytes representing the serialized (on-the-wire) representation of this raw frame.

fn header(&self) -> FrameHeader

Returns a FrameHeader instance corresponding to the headers of the RawFrame.

fn payload(&self) -> &[u8]

Returns a slice representing the payload of the RawFrame.