Struct solicit::http::frame::RawFrame

pub struct RawFrame {
    pub header: FrameHeader,
    pub payload: Vec<u8>,
}

A struct that defines the format of the raw HTTP/2 frame, i.e. the frame as it is read from the wire.

This format is defined in section 4.1. of the HTTP/2 spec.

The RawFrame struct simply stores the raw components of an HTTP/2 frame: its header and the payload as a sequence of bytes.

It does not try to interpret the payload bytes, nor do any validation in terms of its validity based on the frame type given in the header. It is simply a wrapper around the two parts of an HTTP/2 frame.

Fields

header

The parsed header of the frame.

payload

The payload of the frame, as the raw byte sequence, as received on the wire.

Methods

impl RawFrame

fn new(header: FrameHeader) -> RawFrame

Creates a new RawFrame with the given FrameHeader. The payload is left empty.

fn with_payload(header: FrameHeader, payload: Vec<u8>) -> RawFrame

Creates a new RawFrame with the given header and payload.

fn from_buf(buf: &[u8]) -> Option<RawFrame>

Creates a new RawFrame by parsing the given buffer.

Returns

If the first bytes of the buffer represent a valid frame, a RawFrame that represents it is returned. Obviously, the buffer should contain both the header and the payload of the frame.

If the first bytes of the buffer cannot be interpreted as a raw frame, None is returned. This includes the case where the buffer does not contain enough data to contain the entire payload (whose length was advertised in the header).