Module solicit::http::priority

The module exposes an API for defining data prioritization strategies.

Types that implement the DataPrioritizer trait can be used to provide new data for an HttpConnection to send to its peer. Neither the HttpConnection nor the DataPrioritizer have control over exactly when the data is sent. This is left up to the particular client implementations to trigger.

Structs

SimplePrioritizer
An implementation of the DataPrioritizer trait that is based on finding the first stream from the given SessionState instance that can send data and returning this chunk.

Traits

DataPrioritizer
A trait that types that want to provide data to an HTTP/2 connection need to implement.