Module solicit::http::session

Defines the interface for the session-level management of HTTP/2 communication. This is effectively an API that allows hooking into an HTTP/2 connection in order to handle events arising on the connection.

It also provides a default implementation of this interface, the DefaultSession. This implementation is based on keeping a mapping of valid stream IDs to instances of Stream objects. When the session receives a callback for a particular stream ID, it first validates that it represents a valid stream ID and then delegates to the appropriate action of a Stream. This allows clients to easily vary the stream-level logic, without worrying about handling the book-keeping tasks of which streams are active.

Structs

DefaultSession
A simple implementation of the Session trait.

DefaultStream
An implementation of the Stream trait that saves all headers and data in memory.

Traits

Session
A trait that defines methods that need to be defined in order to track the status of a ClientConnection.

Stream
A trait representing a single HTTP/2 client stream. An HTTP/2 connection multiplexes a number of streams.