Trait solicit::http::session::Stream

pub trait Stream {
    fn new(stream_id: StreamId) -> Self;
    fn new_data_chunk(&mut self, data: &[u8]);
    fn set_headers(&mut self, headers: Vec<Header>);
    fn close(&mut self);
    fn id(&self) -> StreamId;
    fn is_closed(&self) -> bool;
}

A trait representing a single HTTP/2 client stream. An HTTP/2 connection multiplexes a number of streams.

The trait defines which operations need to be defined by a type that should be useable as an HTTP/2 stream. By implementing this trait, clients can only implement stream-level logic, such as how the received data should be handled, instead of tracking which streams exist and what their states are.

Required Methods

fn new(stream_id: StreamId) -> Self

Create a new stream with the given ID

fn new_data_chunk(&mut self, data: &[u8])

Handle a new data chunk that has arrived for the stream.

fn set_headers(&mut self, headers: Vec<Header>)

Set headers for a stream. A stream is only allowed to have one set of headers.

fn close(&mut self)

Close the stream.

fn id(&self) -> StreamId

Returns the ID of the stream.

fn is_closed(&self) -> bool

Returns whether the stream is closed.