Struct solicit::http::Header

pub struct Header<'n, 'v> {
    // some fields omitted
}

Represents an HTTP/2 header. Allows both the name and the value to be either an owned or a borrowed byte sequence.

Examples

A new Header can be created by providing an owned or borrowed name, as well as value:

use solicit::http::Header;
// Name and value both borrowed (static) slices.
let h1 = Header::new(b":method", b"GET");
assert_eq!(h1.name(), &b":method"[..]);
assert_eq!(h1.value(), &b"GET"[..]);
// A borrowed slice with a scope-bound lifetime as the value; static name
{
    let value = vec![1];
    let header = Header::new(&b"x-test-head"[..], &value[..]);
    assert_eq!(header.name(), &b"x-test-head"[..]);
    assert_eq!(header.value(), &[1][..]);
}
// An owned value, static name
{
    let value = vec![1];
    let header = Header::new(&b"x-test-head"[..], value);
    assert_eq!(header.name(), &b"x-test-head"[..]);
    assert_eq!(header.value(), &[1][..]);
}
// An owned name, as well as value
{
    let value = vec![1];
    let name = b"x-test-head".to_vec();
    let header = Header::new(name, value);
    assert_eq!(header.name(), &b"x-test-head"[..]);
    assert_eq!(header.value(), &[1][..]);
}

Methods

impl<'n, 'v> Header<'n, 'v>

fn new<N: Into<HeaderPart<'n>>, V: Into<HeaderPart<'v>>>(name: N, value: V) -> Header<'n, 'v>

Creates a new Header with the given name and value.

The name and value need to be convertible into a HeaderPart.

fn name(&self) -> &[u8]

Return a borrowed representation of the Header name.

fn value(&self) -> &[u8]

Return a borrowed representation of the Header value.